backend/requirements.txt
# FastAPI & Server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Document Processing
PyPDF2==3.0.1
python-docx==1.1.0

# Vector Store (In-memory for Windows, ChromaDB for Linux)
# chromadb>=0.5.0  # Uncomment on Linux/Mac with C++ compiler

# HTTP Client for vLLM
httpx==0.25.2

# Utilities
python-dotenv==1.0.0
pydantic==2.5.2
pydantic-settings==2.1.0
